chore: gnoclient: Add Simulate which returns tx events (#4704)
`EstimateGas` simulates a transaction and returns the gas used so that
the caller can estimate the gas fee. But the caller also needs to
estimate the storage fee. This PR adds the more general function
`Simulate` which returns the full deliverTx which contains the estimated
GasUsed as well as the events with storage fee info. (Also simplify
`EstimateGas` to call `Simulate`.) So, the caller can search for storage
deposit events as needed. (We need this in Gnokey Mobile to show the
full estimated transaction fees.)

---------

Signed-off-by: Jeff Thompson <jeff@thefirst.org>
Co-authored-by: n0izn0iz <n0izn0iz@users.noreply.github.com>
Co-authored-by: Morgan <morgan@morganbaz.com>