fix(params): SetBytes should not alter bytes
SetBytes is used by the IBC contract to store the commitment bytes. Then
the relayer submit the proof of existence of those bytes to the counter
party chain, but if the value bytes are different, the proof cannot be
verified.

This PR updates SetBytes to use directly the store instead of passing by
the common `set` method, preventing the value to be aminoJSON encoded.

Removed GetRaw/SetRaw since GetBytes/SetBytes can be used as a
replacement.