feat(commondao): add support for customizing storage in member groups (#4849)
Adds member grouping options support which allows devs to assign a
custom factory function to create custom member storages when a new
member group is created.

These changes are also required for Boards2 realm, to be able to define
custom board member storages, which is important to integrate CommonDAO
with the custom `Permissions` implementation that Boards2 realm
provides.