feat(boards2): add `gno.land/p/gnoland/boards` package (#4821)
The `gno.land/p/gnoland/boards` package is intended to be a part of
Boards2 realm.
It would allow making the realm smaller by defining common types within
the package.
A follow up PR would refactor Boards2 realm to use this package.

Types are kept minimal so specific logic related to saving or deleting
is implemented by the realm.
Package initially had support for deleting and saving but it was removed
in favor of a more light weight and less opinionated boards package.

The original idea of having a generic `Post` type that can represent a
thread, repost or reply is kept within the package.

Repost, flagging and replies support are optional features for `Post`
which could be initialized when needed by the realm that imports boards
package. A storage interface and a default implementations are available
for each one of these.