feat(commondao): improve proposal states (#4737)
This PR adds new proposal states to improve the workflow of proposals:
- **active**: Default state assigned upon creation, where users can vote
or creator can withdraw proposal if there are no votes.
- **passed**: Proposal has been voted and it passes after counting
votes.
- **rejected**: Proposal it's rejected after counting votes.
- **executed**: Proposal passed and has been executed.
- **failed**: Proposal passed but execution failed.
- **withdrawed**: Proposal was withdrawed by the creator, usually
because proposal is not valid anymore or it has to be corrected.

This change moves tally to the *Proposal* type, which allows devs to
tally proposals on render to calculate passed and rejected states, which
can be useful in cases where there is no need to change proposal status
using a TX, for example for proposals that met the deadline and didn't
get enough votes, in that case it would be optional to send a TX to
effectively change status to rejected.

PR also add support for withdrawing proposals when they have no votes.

---------

Co-authored-by: Nemanja Aleksic <kouteki@gmail.com>