feat(commondao)!: change proposal definition tally to receive a readonly member storage (#4754)
This PR removes the `MemberSet` type in favor of the
`ReadonlyMemberStorage`, which is introduced with other grouping related
readonly types.

`MemberSet` was intended as a readonly representation for proposal
definitions to read DAO members. CommonDAO now supports members grouping
which should also be available to the proposal definitions to be able to
tally votes in contexts where grouping is relevant for counting votes,
and specially where direct access to CommonDAO instances is not
possible, for example for proposal definitions defined in third party
realms.

Proposal definition's `Tally()` method now receives a single argument
containing a readonly member storage instead of a member set and the
readonly voting record:
```go
func (ProposalDefinition) Tally(VotingContext) (bool, error)
```

The new `VotingContext` type is defined as:
```go
// VotingContext contains current voting context, which includes a voting
// record for the votes that has been submitted for a proposal and also
// the list of DAO members.
VotingContext struct {
	VotingRecord ReadonlyVotingRecord
	Members      ReadonlyMemberStorage
}
```

Readonly types are introduced to be able for proposal definitions
implemented in third party realms to read relevant DAO information which
might not be accessible directly to these realms.

Member groups can be accessed from the member storage.