refactor: move custom permissions implementation to a sub realm
Having the type defined in a standalone realm is handy to devs looking
to use this implementation as base for their custom permissions
implementation without having to import boards2 realm.

Permissions should ideally be defined in a package but it defines
crossing methods which are not supported right now to be defined within
packages.