feat: crossing and non crossing executable proposal definitions
Support defining executable proposal definitions which could be either
crossing or non crossing. This is required to be able to define proposal
types within pure packages, because crossing methods and functions can
only be defined within realms.

When a proposal definition is defined in a realm it can implement the
crossing execution method, otherwise it must implement non crossing
methods.

Defining non crossing executable proposal types is useful when the
proposal type is defined within a pure package where the definition
execution is defined by a crossing callback for example. Technically it
would also be possible that a proposal definition just updates local
properties on execution, instead of realm state globals.